Add wallet creation methods (#3)
This pull request adds two new components for creating a new wallet: KeystorePage and PassphrasePage. The KeystorePage component generates a random private key and displays it to the user. The PassphrasePage component allows the user to enter a 9-word passphrase to create a new wallet. Both components are accessible from the Create Wallet page.